+++ magda-web-client/src/Components/Dataset/View/DatasetPageDetails.js
@@ -6,6 +6,8 @@ import DistributionRow from "./DistributionRow";
 import queryString from "query-string";
 import defined from "helpers/defined";
 import CommonLink from "Components/Common/CommonLink";
+import DiscourseComments from "Components/Dataset/View/DiscourseComments";
+import { config } from "../../../config";
 import "./DatasetDetails.scss";
 import "./DatasetPageDetails.scss";
 
@@ -80,6 +82,22 @@ class DatasetPageDetails extends Component {
                         )}
                     </div>
                 </div>
+
+                {config.discourseSiteUrl &&
+                config.discourseIntegrationDatasetPage ? (
+                    <div className="row">
+                        <div className="col-sm-12">
+                            <DiscourseComments
+                                title={
+                                    dataset?.title
+                                        ? dataset.title
+                                        : "Untitlted Dataset"
+                                }
+                                datasetId={dataset.identifier}
+                            />
+                        </div>
+                    </div>
+                ) : null}
             </div>
         );
     }